backend/requirements.txt
fastapi
uvicorn
pydantic
httpx
duckdb
pandas
shapely
orjson
tenacity
python-dotenv
rq
redis
pyproj
haversine
pyogrio
pytest
pyarrow
